Improve tab completion of file and directory names
