+++ test/integration-tests/shell.test.ts
@@ -121,172 +121,6 @@ test.describe('Shell', () => {
     });
   });
 
-  test.describe('input tab complete commands', () => {
-    test('should complete ec', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['e', 'c', '\t'])).toEqual('echo ');
-    });
-
-    test('should ignore leading whitespace', async ({ page }) => {
-      expect(await shellInputsSimple(page, [' ', 'e', 'c', '\t'])).toEqual(' echo ');
-    });
-
-    test('should ignore leading whitespace x2', async ({ page }) => {
-      expect(await shellInputsSimple(page, [' ', ' ', 'e', 'c', '\t'])).toEqual('  echo ');
-    });
-
-    test('should show tab completion options', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['e', '\t'])).toMatch(
-        /^e\r\necho {4}env {5}exit {4}export {2}expr\r\n/
-      );
-    });
-
-    test('should do nothing on unknown command', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['u', 'n', 'k', '\t'])).toEqual('unk');
-    });
-
-    test('should arrange in columns', async ({ page }) => {
-      const output = await page.evaluate(async () => {
-        const { shell, output } = await globalThis.cockle.shellSetupEmpty();
-        await shell.setSize(40, 10);
-        await shell.input('t');
-        await shell.input('\t');
-        const ret0 = output.textAndClear();
-
-        await shell.setSize(40, 20);
-        await shell.input('\t');
-        const ret1 = output.textAndClear();
-        return [ret0, ret1];
-      });
-      expect(output[0]).toMatch(/^t\r\ntail\r\ntee\r\ntouch\r\ntr\r\ntree\r\ntty\r\n/);
-      expect(output[1]).toMatch(/^\r\ntail {3}tr\r\ntee {4}tree\r\ntouch {2}tty\r\n/);
-    });
-
-    test('should add common startsWith', async ({ page }) => {
-      const output = await shellInputsSimpleN(page, [['s', 'h', '\t'], ['\t']]);
-      expect(output[0]).toEqual('sha');
-      expect(output[1]).toMatch(/sha1sum {4}sha224sum {2}sha256sum {2}sha384sum {2}sha512sum/);
-    });
-
-    test('should include aliases', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['l', '\t'])).toMatch(
-        /^l\r\nless {5}ll {7}ln {7}logname {2}ls {7}lua\r\n/
-      );
-    });
-
-    test('should complete within a command preserving suffix', async ({ page }) => {
-      const { enter, leftArrow, tab } = keys;
-      const output = await shellInputsSimpleN(page, [['e', 'c', 'X', leftArrow, tab], [enter]]);
-      expect(output[1]).toMatch(/^\r\nX\r\n/);
-    });
-  });
-
-  test.describe('input tab complete filenames', () => {
-    test('should do nothing with unrecognised filename', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['l', 's', ' ', 'z', '\t'])).toEqual('ls z');
-    });
-
-    test('should show tab completion options', async ({ page }) => {
-      const output = await shellInputsSimple(page, ['l', 's', ' ', 'f', 'i', 'l', 'e', '\t']);
-      expect(output).toMatch(/^ls file\r\nfile1 {2}file2\r\n/);
-    });
-
-    test('should add common startsWith', async ({ page }) => {
-      expect(await shellInputsSimple(page, ['l', 's', ' ', 'f', '\t'])).toEqual('ls file');
-    });
-
-    test('should complete single filename, adding trailing space', async ({ page }) => {
-      const output = await shellInputsSimple(page, ['l', 's', ' ', 'f', 'i', 'l', 'e', '1', '\t']);
-      expect(output).toEqual('ls file1 ');
-    });
-
-    test('should complete single directory, adding trailing slash', async ({ page }) => {
-      const output = await shellInputsSimple(page, ['l', 's', ' ', 'd', '\t']);
-      expect(output).toEqual('ls dirA/');
-    });
-
-    test('should list contents if match directory with trailing slash', async ({ page }) => {
-      const output = await shellInputsSimple(page, [
-        'l',
-        's',
-        ' ',
-        '/',
-        'd',
-        'r',
-        'i',
-        'v',
-        'e',
-        '/',
-        '\t'
-      ]);
-      expect(output).toMatch(/^ls \/drive\/\r\ndirA\/ {2}file1 {2}file2\r\n/);
-      expect(output).toMatch(/ls \/drive\/$/);
-    });
-
-    test('should list contents if match directory without trailing slash', async ({ page }) => {
-      const output = await shellInputsSimple(page, [
-        'l',
-        's',
-        ' ',
-        '/',
-        'd',
-        'r',
-        'i',
-        'v',
-        'e',
-        '\t'
-      ]);
-      expect(output).toMatch(/^ls \/drive\r\ndirA\/ {2}file1 {2}file2\r\n/);
-      expect(output).toMatch(/ls \/drive\/$/);
-    });
-
-    test('should support . for current directory', async ({ page }) => {
-      const output = await shellInputsSimple(page, ['l', 's', ' ', '.', '\t']);
-      expect(output).toMatch(/^ls .\r\n..\/ {2}.\/\r\n/);
-      expect(output).toMatch(/ls .$/);
-    });
-
-    test('should support . for current directory 2', async ({ page }) => {
-      const output = await shellInputsSimple(page, [
-        'l',
-        's',
-        ' ',
-        '.',
-        '/',
-        'f',
-        'i',
-        'l',
-        'e',
-        '\t'
-      ]);
-      expect(output).toMatch(/^ls .\/file\r\nfile1 {2}file2\r\n/);
-      expect(output).toMatch(/ls .\/file$/);
-    });
-
-    test('should support .. for parent directory', async ({ page }) => {
-      const output = await shellInputsSimple(page, ['l', 's', ' ', '.', '.', '/', 'd', 'r', '\t']);
-      expect(output).toEqual('ls ../drive/');
-    });
-
-    test('should show dot files/directories', async ({ page }) => {
-      const options = {
-        initialDirectories: ['.adir'],
-        initialFiles: { '.afile1': '', '.afile2': '' }
-      };
-      const output = await shellInputsSimpleN(
-        page,
-        [
-          ['l', 's', ' ', '.', '\t'],
-          ['l', 's', ' ', '.', 'a', '\t'],
-          ['l', 's', ' ', '.', 'a', 'f', '\t']
-        ],
-        options
-      );
-      expect(output[0]).toMatch(/^ls .\r\n..\/ {6}.\/ {7}.adir\/ {3}.afile1 {2}.afile2\r\n/);
-      expect(output[1]).toMatch(/^ls .a\r\n.adir\/ {3}.afile1 {2}.afile2\r\n/);
-      expect(output[2]).toEqual('ls .afile');
-    });
-  });
-
   test.describe('setSize', () => {
     test('should set envVars', async ({ page }) => {
       const output = await page.evaluate(async () => {
